MinGW build fixes
Ref: ba70ecf7593e20eebae1a5249a7cdda9af39e1b1

+++ src/gui/qgsattributeformlegacyinterface.cpp
@@ -33,7 +33,7 @@ QgsAttributeFormLegacyInterface::QgsAttributeFormLegacyInterface( const QString&
 
   QString initLayer = QString( "%1 = sip.wrapinstance( %2, qgis.core.QgsVectorLayer )" )
                       .arg( mPyLayerVarName )
-                      .arg(( unsigned long ) form->layer() );
+                      .arg(( intptr_t ) form->layer() );
 
   QgsPythonRunner::run( initLayer );
 }
@@ -61,7 +61,7 @@ void QgsAttributeFormLegacyInterface::featureChanged()
   QString pyFeatureVarName = QString( "_qgis_feature_%1" ).arg( dt.toString( "yyyyMMddhhmmsszzz" ) );
   QString initFeature = QString( "%1 = sip.wrapinstance( %2, qgis.core.QgsFeature )" )
                         .arg( pyFeatureVarName )
-                        .arg(( unsigned long ) & form()->feature() );
+                        .arg(( intptr_t ) & form()->feature() );
 
   QgsPythonRunner::run( initFeature );
